requirements.txt
streamlit==1.29.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.2
Pillow==10.1.0
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0